templates/project/python/pyproject.toml
[project]
name = "{{PROJECT_NAME}}"
version = "0.1.0"
description = "FastAPI + SQLAlchemy + Lakebase database branches"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "psycopg[binary]>=3.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
